Add dll and windows compliance with update

+++ Src/Entity/Bomb/Bomb.cpp
@@ -62,7 +62,7 @@ void is::Bomb::explode()
 
 void is::Bomb::timer()
 {
-	ssize_t now = time(nullptr);
+	long int now = time(nullptr);
 
 	if (_isPaused && now > _lastTime) {
 		_time++;
